Don't acquire known-unusable locks in LockingService
The DynamoDB lock table is now one of the main costs of a reindex in the
catalogue pipeline, so I've been looking for ways to make it more
efficient.  We do see locking failures in the matcher logs, so some
processes are trying to acquire the same/similar locks.

Previously, the locking service would lock IDs in sequence, e.g.:

    lockDao.lock(id1)
    lockDao.lock(id2)
    lockDao.lock(id3)

But if we get a failure on locking `id1`, trying to lock `id2` and `id3`
is pointless.  We create and then delete two locks that we know we can
never use, which is expensive and potentially stops another process from
locking one of those IDs.

Now, if the locking service detects a failure to lock `id1`, it skips locking
`id2` and `id3`, and goes straight to the cleanup stage.

I don't know how much of a difference this will make -- I'd bet a few
percentage points at best -- because this code is already fairly heavily
optimised, but having spotted this simple optimisation I figured it
couldn't hurt to add.