Inspect queue size based on metrics, not waiting for messages to appear
Waiting for messages to appear isn't ideal: normally when we're using an
SQS queue in a test, we also have a worker process reading from the queue.

If the fixture doesn't pick up any messages from the queue, that doesn't
tell us anything – there might be a message in-flight with the worker
we're testing.  We already inspect the queue metrics to see if there's
a message in flight, so let's just do that and skip the bit where we
wait for the message to be processed out.

The waiting is normally handled by the `eventually {…}`, with the
advantage that it stops as soon as it's done rather than waiting N seconds.