Remove the MessageTransform trait
We never use the optional context which is added in `callTransform()`,
so removing it allows us to shed a bunch of type parameters.  I've also
tried to do away with the word "transform" in this context, which is an
overloaded term, and replace it with "parsing" messages we get from the
queue.  Hopefully this makes the code easier to follow.