Don't return a tuple from MessageTransform.callTransform
Previously this returned a tuple with the context as the second item,
but we no longer use this when we get the value in Worker.  Let's simplify
MessageTransform by ditching the second argument, which allows us to
remove a bunch of code and type parameters we no longer need.

(Or to be more correct, were never anything except an `Instant`.)