Change the behaviour of failed SQS messages
Previously, a failed message would be deleted from the queue.  This patch
modifies that behaviour: now we delete the message from the queue *and*
forward it to the DLQ.  This gives us slightly more visibility on the
failure, and the ability to redrive it if we know it's a transient error.