Allow retrying some terminal failures, if they're AWS flaking out
This is catch-all logic for any unexpected failures that we don't catch
elsewhere, e.g. the SNS sending error we saw last week -- the sort of
"once in a blue moon" error we don't see very often, but still want to be
able to retry.

It doesn't replace the retrying logic elsewhere in the stack, rather adds
SQS retrying on top of it, when everything else has failed.  e.g. if we're
doing 1000 S3 operations per SQS message, we want retrying on those
operations individually, or we might never succeed.  This is meant to
catch errors we can't or haven't handled elsewhere.

Actually terminal failures (e.g. publishing to a non-existent SNS topic)
will still be treated as terminal, and see the message deleted from the
queue.