Retry looking up the destination blob when replicating to Azure
Yesterday we got a 409 error from the bag replicator:

> BlobImmutableDueToLegalHold: This operation is not permitted as the
> blob is immutable due to one or more legal holds.

This should be impossible -- the Azure replicator is meant to check for
an existing object in the destination, and skip overwriting it if present.
(If the object matches what it was about to write, it's a no-op.  If it's
different, it gets flagged as an error.)

Upon reading the code, we spotted a bug in the Azure logic: if there was
*any* error retrieving the existing blob, it would proceed to overwrite
the existing object.  This includes transient errors, e.g. a timeout.

This makes two changes to this logic:

*   It uses the AzureStreamStore and S3StreamStore, which both have code
    to retry flaky errors.  We already use this in S3Transfer; it should
    prevent transient errors breaking the Azure replication.

*   If there's any error looking up the destination stream which isn't
    a 404 Not Found, it logs a warning, so we can debug it next time.