Pass a new type around inside AzureTransfer
To copy an object from S3 to Azure, we need to know three things:

* the bucket
* the key
* the size

In the V1 SDK, we could use `S3ObjectSummary`, which contains all three
values.  In the V2 SDK, this has been replaced with `S3Object`, which is
missing bucket.

We also have our own `S3ObjectLocation` type, but this is inadequate:

                     | bucket |  key   |  size
---------------------+--------+--------+--------
S3ObjectSummary (V1) |   ✓    |   ✓    |   ✓
S3Object (V2)        |   ✗    |   ✓    |   ✓
S3ObjectLocation     |   ✓    |   ✓    |   ✗

As part of the prep for the V2 SDK, this introduces a new `SourceS3Object`
type which can replace `S3ObjectSummary` inside the Azure transfer classes.